Rename `lib_regions_of_interes` into `roi`
