+++ fractal_tasks_core/tasks/apply_registration_to_ROI_tables.py
@@ -25,7 +25,7 @@ import zarr
 from pydantic.decorator import validate_arguments
 
 from fractal_tasks_core.lib_ngff import load_NgffWellMeta
-from fractal_tasks_core.lib_regions_of_interest import (
+from fractal_tasks_core.roi import (
     are_ROI_table_columns_valid,
 )
 from fractal_tasks_core.tables import write_table
