+++ fractal_tasks_core/tasks/calculate_registration_image_based.py
@@ -28,14 +28,14 @@ from skimage.registration import phase_cross_correlation
 from fractal_tasks_core.channels import get_channel_from_image_zarr
 from fractal_tasks_core.channels import OmeroChannel
 from fractal_tasks_core.lib_ngff import load_NgffImageMeta
-from fractal_tasks_core.lib_regions_of_interest import check_valid_ROI_indices
-from fractal_tasks_core.lib_regions_of_interest import (
+from fractal_tasks_core.roi import check_valid_ROI_indices
+from fractal_tasks_core.roi import (
     convert_indices_to_regions,
 )
-from fractal_tasks_core.lib_regions_of_interest import (
+from fractal_tasks_core.roi import (
     convert_ROI_table_to_indices,
 )
-from fractal_tasks_core.lib_regions_of_interest import load_region
+from fractal_tasks_core.roi import load_region
 from fractal_tasks_core.tables import write_table
 
 logger = logging.getLogger(__name__)
