+++ fractal_tasks_core/tasks/cellpose_segmentation.py
@@ -38,22 +38,22 @@ from fractal_tasks_core.lib_input_models import Channel
 from fractal_tasks_core.lib_masked_loading import masked_loading_wrapper
 from fractal_tasks_core.lib_ngff import load_NgffImageMeta
 from fractal_tasks_core.lib_pyramid_creation import build_pyramid
-from fractal_tasks_core.lib_regions_of_interest import (
-    array_to_bounding_box_table,
-)
-from fractal_tasks_core.lib_regions_of_interest import check_valid_ROI_indices
-from fractal_tasks_core.lib_regions_of_interest import (
-    convert_ROI_table_to_indices,
-)
-from fractal_tasks_core.lib_regions_of_interest import empty_bounding_box_table
-from fractal_tasks_core.lib_regions_of_interest import (
-    find_overlaps_in_ROI_indices,
-)
-from fractal_tasks_core.lib_regions_of_interest import get_overlapping_pairs_3D
-from fractal_tasks_core.lib_regions_of_interest import is_ROI_table_valid
-from fractal_tasks_core.lib_regions_of_interest import load_region
 from fractal_tasks_core.lib_write import prepare_label_group
 from fractal_tasks_core.lib_zattrs_utils import rescale_datasets
+from fractal_tasks_core.roi import (
+    array_to_bounding_box_table,
+)
+from fractal_tasks_core.roi import check_valid_ROI_indices
+from fractal_tasks_core.roi import (
+    convert_ROI_table_to_indices,
+)
+from fractal_tasks_core.roi import empty_bounding_box_table
+from fractal_tasks_core.roi import (
+    find_overlaps_in_ROI_indices,
+)
+from fractal_tasks_core.roi import get_overlapping_pairs_3D
+from fractal_tasks_core.roi import is_ROI_table_valid
+from fractal_tasks_core.roi import load_region
 from fractal_tasks_core.tables import write_table
 
 logger = logging.getLogger(__name__)
