+++ fractal_tasks_core/tasks/copy_ome_zarr.py
@@ -23,10 +23,10 @@ from pydantic.decorator import validate_arguments
 
 import fractal_tasks_core
 from fractal_tasks_core.lib_ngff import load_NgffImageMeta
-from fractal_tasks_core.lib_regions_of_interest import (
+from fractal_tasks_core.lib_write import open_zarr_group_with_overwrite
+from fractal_tasks_core.roi import (
     convert_ROIs_from_3D_to_2D,
 )
-from fractal_tasks_core.lib_write import open_zarr_group_with_overwrite
 from fractal_tasks_core.tables import write_table
 
 logger = logging.getLogger(__name__)
