+++ fractal_tasks_core/tasks/illumination_correction.py
@@ -31,8 +31,8 @@ from fractal_tasks_core.channels import get_omero_channel_list
 from fractal_tasks_core.channels import OmeroChannel
 from fractal_tasks_core.lib_ngff import load_NgffImageMeta
 from fractal_tasks_core.lib_pyramid_creation import build_pyramid
-from fractal_tasks_core.lib_regions_of_interest import check_valid_ROI_indices
-from fractal_tasks_core.lib_regions_of_interest import (
+from fractal_tasks_core.roi import check_valid_ROI_indices
+from fractal_tasks_core.roi import (
     convert_ROI_table_to_indices,
 )
