+++ tests/tasks/test_unit_illumination_correction.py
@@ -11,7 +11,7 @@ from pytest import LogCaptureFixture
 from pytest import MonkeyPatch
 
 from fractal_tasks_core.lib_ngff import load_NgffImageMeta
-from fractal_tasks_core.lib_regions_of_interest import (
+from fractal_tasks_core.roi import (
     convert_ROI_table_to_indices,
 )
 from fractal_tasks_core.tasks.illumination_correction import correct
